Move ExplorationTechnique base class to own file
