+++ angr/exploration_techniques/loop_seer.py
@@ -1,10 +1,9 @@
 from __future__ import annotations
 import logging
 
-from . import ExplorationTechnique
 from angr.knowledge_base import KnowledgeBase
 from angr.knowledge_plugins.functions import Function
-
+from .base import ExplorationTechnique
 
 l = logging.getLogger(name=__name__)
 
@@ -52,6 +51,10 @@ class LoopSeer(ExplorationTechnique):
         self.limit_concrete_loops = limit_concrete_loops
         self.loops = {}
         self.cut_succs = []
+
+        # Delayed import
+        from angr.analyses.loopfinder import Loop
+
         if type(loops) is Loop:
             loops = [loops]
 
@@ -231,6 +234,3 @@ class LoopSeer(ExplorationTechnique):
             f = func
 
         return f
-
-
-from angr.analyses.loopfinder import Loop
