+++ angr/exploration_techniques/manual_mergepoint.py
@@ -1,7 +1,8 @@
 from __future__ import annotations
+
 import logging
 
-from . import ExplorationTechnique
+from .base import ExplorationTechnique
 
 l = logging.getLogger(name=__name__)
