+++ angr/exploration_techniques/slicecutor.py
@@ -1,8 +1,10 @@
 from __future__ import annotations
-from angr.errors import AngrExitError
-from . import ExplorationTechnique
 
 import logging
+
+from angr.errors import AngrExitError
+from .base import ExplorationTechnique
+
 
 l = logging.getLogger(name=__name__)
