+++ angr/exploration_techniques/spiller.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 import contextlib
 import logging
 
-from . import ExplorationTechnique
+from .base import ExplorationTechnique
 
 
 l = logging.getLogger(name=__name__)
