+++ angr/exploration_techniques/suggestions.py
@@ -1,12 +1,14 @@
 from __future__ import annotations
+
 import logging
+
 import claripy
 
-from . import ExplorationTechnique
 from angr.misc.ux import once
 from angr.misc.picklable_lock import PicklableLock
 from angr.state_plugins.sim_action import SimActionConstraint
 from angr.state_plugins.sim_action_object import SimActionObject
+from .base import ExplorationTechnique
 
 l = logging.getLogger(__name__)
